Bugfix: wrong prefix boundary when compressing dataset.
This bug was also present in jLCM, and it was detected by
M.Kirchgessner. It was not causing wrong results, but a small
performance drop on big datasets.
We were comparing to the core_item (extension item) in order to
fin the prefix boundary, in the compression algorithm,
although the items were renumbered previously.
Counters::maxCandidate was an appropriate replacement.